+++ app/extensions/safe/api/helpers.js
@@ -2,6 +2,7 @@ import { CONFIG } from 'appConstants';
 
 const crypto = require('crypto'); // electron deps will be available inside browser
 const { Readable } = require('stream');
+const errConst = require('../err-constants');
 
 const handles = new Map();
 
@@ -41,7 +42,7 @@ export const getObj = (handle, supportNull) => new Promise((resolve, reject) => 
   if (obj) {
     return resolve(obj);
   }
-  return reject(new Error(`Invalid handle: ${handle}`));
+  return reject(new Error(errConst.INVALID_HANDLE.msg(handle)));
 });
 
 // this helper functions will free all object associated with app\